+++ .github/workflows/pythonapp.yml
@@ -21,6 +21,8 @@ jobs:
   lint:
     name: Lint code
     runs-on: ubuntu-latest
+    env:
+      UFLX_ENABLE_TESTING: "1"
     steps:
       - name: Set up Python
         uses: actions/setup-python@v5
@@ -44,6 +46,8 @@ jobs:
       matrix:
         os: [ubuntu-24.04]
         python-version: ["3.11", "3.12", "3.13", "3.14"]
+    env:
+      UFLX_ENABLE_TESTING: "1"
     steps:
       - uses: actions/checkout@v4
       - name: Set up Python
