+++ test/test_algorithms.py
@@ -1,13 +1,15 @@
 """Test graph algorithms."""
 
-from utils import LagrangeElement, triangle
 
 from uflx import TestFunction, TrialFunction, coordinate_element, dx, function_space
 from uflx.graphs.algorithms import replace
+from uflx.test_utils import triangle, LagrangeElement
+
 
 
 def test_replace():
     """Test replace algorithm."""
+
     element = LagrangeElement(triangle, 2)
     domain = coordinate_element(LagrangeElement(triangle, 1, (2,)))
     space = function_space(domain, element)
